requirements.txt
#requirements.txt
snowflake-connector-python
streamlit-folium
folium
Python-Geohash